+++ R/cansim_vectors.R
@@ -208,7 +208,7 @@ get_cansim_vector<-function(vectors, start_time = as.Date("1800-01-01"), end_tim
       }
       if (is.null(response)) return(response)
       if (response$status_code!=200) {
-        stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response))
+        stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response),call.=FALSE)
       }
       data <- httr::content(response)
       data1 <- Filter(function(x)x$status=="SUCCESS",data)
@@ -318,7 +318,7 @@ get_cansim_vector_for_latest_periods<-function(vectors, periods=NULL,
       response <- post_with_timeout_retry(url, body=vectors_string, timeout = timeout)
       if (is.null(response)) return(response)
       if (response$status_code!=200) {
-        stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response))
+        stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response),call.=FALSE)
       }
       data <- httr::content(response)
       data1 <- Filter(function(x)x$status=="SUCCESS",data)
@@ -385,7 +385,7 @@ get_cansim_vector_for_latest_periods<-function(vectors, periods=NULL,
 #' @param periods Optional numeric value for number of latest periods to retrieve data for, default is \code{NULL} in which case data for all periods is downloaded.
 #' Alternatively this can be specified by
 #' coordinate if tableCoordinates is a data frame, this argument will be ignored if that data frame as a "periods" column.
-#' @param language \code{"en"} or \code{"english"} for English and \code{"fr"} or \code{"french"} for French language versions (defaults to English)
+#' @param language \code{"eng"} or \code{"english"} for English and \code{"fra"} or \code{"french"} for French language versions (defaults to English)
 #' @param refresh (Optional) When set to \code{TRUE}, forces a reload of data table (default is \code{FALSE})
 #' @param timeout (Optional) Timeout in seconds for downloading cansim table to work around scenarios where StatCan servers drop the network connection.
 #' @param factors (Optional) Logical value indicating if dimensions should be converted to factors. (Default set to \code{TRUE}).
@@ -458,7 +458,7 @@ get_cansim_data_for_table_coord_periods<-function(tableCoordinates, periods=NULL
       message(paste0("Accessing CANSIM NDM coordinates from Statistics Canada",addition))
       response <- post_with_timeout_retry(url, body=body_string, timeout = timeout)
       if (response$status_code!=200) {
-        stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response))
+        stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response),call.=FALSE)
       }
       data <- httr::content(response)
       data1 <- Filter(function(x)x$status=="SUCCESS",data)
@@ -565,7 +565,7 @@ get_cansim_vector_info <- function(vectors){
   vectors_string=paste0("[",paste(purrr::map(as.character(vectors),function(x)paste0('{"vectorId":',x,'}')),collapse = ", "),"]")
   response <- post_with_timeout_retry(url, body=vectors_string)
   if (response$status_code!=200) {
-    stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response))
+    stop("Problem downloading data, status code ",response$status_code,"\n",httr::content(response),call.=FALSE)
   }
   data <- httr::content(response)
   data1 <- Filter(function(x)x$status=="SUCCESS",data)